requirements.txt
aiohttp
discord.py
Pillow